api/requirements.txt
## Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
numpy==2.1.0
Pillow==10.4.0

## Computer Vision
opencv-python-headless==4.10.0.84
mediapipe==0.10.18

## Background Removal
rembg[gpu]==2.0.59
onnxruntime==1.19.0

## Image Enhancement (optional — graceful fallback if not installed)
# realesrgan>=0.3.0        # Super-resolution (requires basicsr)
# basicsr>=1.4.2           # Real-ESRGAN dependency
# gfpgan>=1.3.8            # Face restoration

## Skin Analysis v2
scikit-image>=0.22.0
